Refactor based on feedback. Added StorageUtils.
